# Add subdaily time binning to VWC workflow