## Summary

This PR introduces a configurable time-binning system to produce soil moisture estimates at a subdaily temporal resolution.

**New Parameters:**
- `-bin_hours`: Time bin size (Valid bin hours: 1, 2, 3, 4, 6, 8, 12, 24 - divisors of 24)
- `-minvalperbin`: Minimum tracks per bin (replaces `-minvalperday` for subdaily)
- `-bin_offset`: Bin timing offset - how many hours after midnight does the first bin start? (0 ≤ offset < bin_hours)

**Backwards Compatibility:**
- `-minvalperday` still supported for daily (24hr) analysis with deprecation warning

## Example Workflow

```bash
# --- Daily Analysis (unchanged behavior) ---
vwc_input p038 2024                              # Configure with defaults
vwc p038 2024                                    # 24-hour daily binning, same as current daily estimates

# --- 6-hour Subdaily Analysis ---  
vwc_input p038 2024 -bin_hours 6 -minvalperbin 5 # Configure subdaily parameters
vwc p038 2024                                    # Use JSON defaults
vwc p038 2024 -bin_hours 12 -minvalperbin 8       # Override JSON parameters

# --- Offset Binning (shifted timing) ---
vwc p038 2024 -bin_hours 6 -bin_offset 2         # 02-08h, 08-14h, 14-20h, 20-02h bins
vwc p038 2024 -bin_hours 8 -bin_offset 4         # 04-12h, 12-20h, 20-04h bins

# --- Mixed Temporal Resolutions ---
vwc p038 2024 -bin_hours 24                      # Creates daily files  
vwc p038 2024 -bin_hours 6                       # Creates separate 6hr files
vwc p038 2024 -bin_hours 12                      # Creates separate 12hr files
```

**File Outputs:** All output files now include a new suffix with frequency (even L2) + bin length + offset. 

```
Files/p038/
├── p038_phase_L2_24hr+0.txt   # Daily with zero offset
├── p038_phase_L2_6hr+0.txt    # 6-hour subdaily, zero offset  
├── p038_phase_L2_6hr+2.txt    # 6-hour subdaily, 2-hour offset
├── p038_vwc_L2_24hr+0.txt     # Daily VWC
├── p038_vwc_L2_6hr+0.txt      # 6-hour VWC, zero offset
├── p038_vwc_L2_6hr+2.txt      # 6-hour VWC, 2-hour offset
└── [corresponding plots also with new temporal resolution suffixes]
```

**File Format Example:** Now always includes frequency used in the header. Note that there was an existing placeholder(?) column "empty" that I didn't touch to keep the format backwards compatible.

Daily format (8 columns, unchanged)
```
% L2 phase results for station P038
% Year DOY   Ph    Phsig NormA  empty  Mon Day 
 2024   1   8.96   5.39  1.011  0.00    1  1 
 2024   2   7.70   4.47  0.987  0.00    1  2 
```
```
% Soil Moisture Results for GNSS Station p038 
% Frequency used: L2C 
% https://github.com/kristinemlarson/gnssrefl 
% FracYr    Year   DOY   VWC Month Day 
 2024.0027 2024    1    0.075   1   1 
 2024.0055 2024    2    0.057   1   2 
```

Subdaily format (9 columns, adds BinStart)  
```
% L2 phase results for station P038
% Subdaily phase averaging with 6-hour bins (offset: 0h)
% Bin schedule: Bin0: 00-06h  Bin1: 06-12h  Bin2: 12-18h  Bin3: 18-24h
% Year DOY   Ph    Phsig NormA  empty  Mon Day BinStart
 2024   1   7.21   6.68  1.019  0.00    1  1    0
 2024   1  11.39   3.54  1.037  0.00    1  1    6
 2024   1   7.71   5.22  0.998  0.00    1  1   12
 2024   1   8.73   4.58  0.972  0.00    1  1   18
 2024   2   6.97   5.49  1.006  0.00    1  2    0
```

```
% Soil Moisture Results for GNSS Station p038 
% Frequency used: L2C 
% https://github.com/kristinemlarson/gnssrefl 
% Subdaily VWC with 6-hour bins (offset: 0h)
% Bin schedule: Bin0: 00-06h  Bin1: 06-12h  Bin2: 12-18h  Bin3: 18-24h
% FracYr    Year   DOY   VWC Month Day BinStart 
 2024.0027 2024    1    0.056   1   1   0 
 2024.0027 2024    1    0.118   1   1   6 
```

## Example Results

Here is a comparison of daily estimates vs 6 hour bins from the P038 site in 2024: 

<img width="3559" height="2906" alt="P038_vwc_comparison_6hr" src="https://github.com/user-attachments/assets/7f841272-a8dd-4924-996f-14725cc9de53">

By running vwc with 6 hour bins for every possible offset value (0-5) and combining the results we can achieve hourly soil moisture estimates with a rolling 6 hour bin. Here is hourly measurements over December 2024 at P038:

<img width="3705" height="1836" alt="p038_vwc_timeseries_2024-12-01_to_2024-12-31" src="https://github.com/user-attachments/assets/b1f60eac-9978-47d1-82b9-aa65e94e7bdd" />

## Breaking Changes

The file naming convention for **all outputs from vwc has changed** to ensure uniqueness between different temporal resolutions/bin offsets.

* **Old Format:** `p038_phase.txt`
* **New Format:** `p038_phase_L2_24hr+0.txt`

Any downstream scripts that parse filenames will need to be updated to accommodate the new `_{bin_duration}hr+{offset}` suffix.