@GeorgeTownsendd 2025-08-19
I've just added an addition commit (#6475775) which adds the "vwc_hourly" command to create an hourly vwc file with a rolling bin. It can be used like this:

```
vwc_hourly p038 2024                            # 6 hour rolling bin, default
vwc_hourly p038 2024 -bin_hours 12              # 12 hour rolling bin
```

You will need minvalperbin to be relatively low to get uninterrupted hourly data. I set 5 as the default. 

The command will run the ```vwc``` workflow for each available offset, then combine the output files. It will also run a 24 hour ```vwc``` workflow for comparison. 

The results file looks like this at ```$REFL_CODE/Files/p038/p038_vwc_L2_rolling12hr.txt```: 

```
% Soil Moisture Results for GNSS Station p038
% Frequency used: L2C
% https://github.com/kristinemlarson/gnssrefl
% Hourly rolling VWC from 12-hour windows
% FracYr    Year   DOY   VWC Month Day BinStart
 2024.0027 2024    1    0.104   1   1   0
 2024.0027 2024    1    0.106   1   1   1
 2024.0027 2024    1    0.103   1   1   2
 2024.0027 2024    1    0.111   1   1   3
 2024.0027 2024    1    0.117   1   1   4
 2024.0027 2024    1    0.113   1   1   5
 2024.0027 2024    1    0.107   1   1   6
 2024.0027 2024    1    0.095   1   1   7
 2024.0027 2024    1    0.089   1   1   8
 2024.0027 2024    1    0.098   1   1   9
 2024.0027 2024    1    0.090   1   1  10
 2024.0027 2024    1    0.090   1   1  11
 2024.0027 2024    1    0.085   1   1  12
 2024.0027 2024    1    0.076   1   1  13
 2024.0027 2024    1    0.079   1   1  14
 2024.0027 2024    1    0.070   1   1  15
 2024.0027 2024    1    0.065   1   1  16
 2024.0027 2024    1    0.073   1   1  17
 2024.0027 2024    1    0.074   1   1  18
 2024.0027 2024    1    0.086   1   1  19
 2024.0027 2024    1    0.091   1   1  20
 2024.0027 2024    1    0.080   1   1  21
 2024.0027 2024    1    0.083   1   1  22
 2024.0027 2024    1    0.080   1   1  23
 2024.0055 2024    2    0.078   1   2   0
 2024.0055 2024    2    0.077   1   2   1
```

It will suppress the plots from ```vwc``` but display a single new plot: 

<img width="1135" height="597" alt="gnssrefl_hourly_output" src="https://github.com/user-attachments/assets/25daa9cb-7b0d-4892-b9fd-b246646843c9" />

(the plot is actually useful when you have an interactive plot to pan/zoom)