@GeorgeTownsendd 2025-08-20
An issue I've noticed with my current implementation: each offset file has a different vegetation correction applied due to different leveling baselines. Thus, there is a systematic bias between each offset file. 

I will work on a way to apply a consistent vegetation to the hourly data and submit a fix, when I have time over the next days/weeks. 

Also happy to take any other feedback for additional changes.