@kristinemlarson 2025-08-20
should i accept what you have submitted so far?  i meant to do it yesterday.

i think we did 12 hour solutions at one point - but we only had  - at most
- 8 L2C satellites by then.
now with 24+, i imagine it is way more robust.  i have meant to share our
high vegetation model -
but i never successfully ported it because i hadn't kept some files.

k.


On Wed, Aug 20, 2025 at 6:43 AM George Townsend ***@***.***>
wrote:

> *GeorgeTownsendd* left a comment (kristinemlarson/gnssrefl#357)
> <https://github.com/kristinemlarson/gnssrefl/pull/357#issuecomment-3206399026>
>
> An issue I've noticed with my current implementation: each offset file has
> a different vegetation correction applied due to different leveling
> baselines. Thus, there is a systematic bias between each offset file.
>
> I will work on a way to apply a consistent vegetation to the hourly data
> and submit a fix, when I have time over the next days/weeks.
>
> Also happy to take any other feedback for additional changes.
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/kristinemlarson/gnssrefl/pull/357#issuecomment-3206399026>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AFSDAPHSQ325PS7XOVYSW733OR3OXAVCNFSM6AAAAACED4UFXCVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZTEMBWGM4TSMBSGY>
> .
> You are receiving this because you are subscribed to this thread.Message
> ID: ***@***.***>
>


-- 
Kristine M. Larson
***@***.***
https://www.kristinelarson.net <http://www.kristinelarson.net>
https://gnss-reflections.org
https://github.com/kristinemlarson

Have you tried reading the documentation?