@kristinemlarson 2025-08-21
okey doke! i will upload my current version (was working on something for
the people
that wanted  midnite crossing arcs) and then will accept yours.

and then i will point you to my high veg efforts.  that info i will email
you privately.

k.

On Thu, Aug 21, 2025 at 1:28 AM George Townsend ***@***.***>
wrote:

> *GeorgeTownsendd* left a comment (kristinemlarson/gnssrefl#357)
> <https://github.com/kristinemlarson/gnssrefl/pull/357#issuecomment-3209537249>
>
> I think it could be merged now, but it is up to you.
>
> As it is, the code is working and useful to start experimenting with. I
> think the bias is minor relative to actual diurnal changes.
>
> In case you do merge, I made issue #358
> <https://github.com/kristinemlarson/gnssrefl/issues/358> for posterity
> and just made an update to the logging in the code to directly link that
> github issue in the console when a user runs vwc_hourly.
>
> If you find any bits of code from the high vegetation model please do
> share and I may be able to help in porting it! :)
>
> Also please feel free to let me know any preferences you have for future
> contributions.
>
> Cheers,
> George
>
> —
> Reply to this email directly, view it on GitHub
> <https://github.com/kristinemlarson/gnssrefl/pull/357#issuecomment-3209537249>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AFSDAPF3BH4NDT2EJJ4YIK33OV7LVAVCNFSM6AAAAACED4UFXCVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZTEMBZGUZTOMRUHE>
> .
> You are receiving this because you commented.Message ID:
> ***@***.***>
>


-- 
Kristine M. Larson
***@***.***
https://www.kristinelarson.net <http://www.kristinelarson.net>
https://gnss-reflections.org
https://github.com/kristinemlarson

Have you tried reading the documentation?